Support config reloads, improve cache refresh behavior
External labels may be updated in Prometheus at runtime, which affects
the series exported to GCM. The underlying series cache thus needs to be
invalidated when the external labels change.

The periodic cache invalidation (to respond to changes in metric
metadata) now is spread evenly over time to avoid load spikes.

b/175105025

Change-Id: I518e0038cbd6de3f3f6c6f6134e2e1cb08487e94