Implement series cache garbage collection
Periodically remove entries from the cache which haven't been updated in
a while. Sample interval tracking is removed as it serves no purpose as
the API resolves cumulative time overlaps by replacing the existing overlapping value.
It shouldn't happen to begin with, but we don't need client side
tracking to avoid repeated write failures if it does.

b/175287427
b/175379839

Change-Id: Iee7a1e7260d7d3c2b3278624ccaf4b1bc01f112a