Use GCM metric type suffix, fix _sum kind for summaries
GPE metrics are suffixed with the Prometheus type to avoid cases where
metrics with the same name but different types cause write conflicts.

Change the GCM kind of the _sum metric of Prometheus summaries to gauges
as summarized elements may be negative, causing the sum to go up or down
over time.

b/175550361
b/175379682

Change-Id: I13c444fbf687432e98b7b241c58515dd7ef4089c