Switch to controller-runtime logging approach
The controller-runtime package comes with its own logging interface. The
go-kit logger cannot be plugged into it properly due to semantic differences.
Thus for the operator binary, we commit to controller-runtimes logging
approach for overall simplicity and consistency.

Logging of the webhook webserver is temporarily disbaled in this PR as
we'll switch to controller-runtime's webhook server in a follow-up.