Dual-wirte unknown metrics, write summary _sum as counter
Write unknown-typed Prometheus metrics as two GCM series, a gauge and a
cumulative so the right choice can be picked at query time.

Write summary _sum series as counter as well as they are in the vast
majority of cases. Use the same suffix-extension as for unknown counters
as the plain foo_sum/summary is already used for the gauge type for
existing users.
In virtually all cases a counter should be the right choice. But in the
future we could also dual write the summary sum if use cases for
negative observations appear.