Scrape port numbers even if not specified on Pod
If a port number is not specified on the Pod resource our previous
filtering mechanism didn't generate a target.
However, we can safely produce targets in this scenario in any case
as we generate a scrape config per specified endpoint and can simply
hardcode all target candidates to the given port number.