Support ungrouped histograms
The exposition format allows for series for a single histogram to not be
grouped together. It only requires all histograms for a metric name to
be grouped together and the buckets for a single histogram to be in
ascending order within that group.

While processing a batch we build a cache for each histogram and produce
a GCM distribution once we've observed sum, count, and infinity bucket
series.