Fix ClusterRules and Rules by location
We previously incorrectly assumed we could label data in regional
cluster by the zone of the machine it originated from. That may
cause theoretical collisions however and we must always set location
to the cluster's official location.
This in practice makes our lives easier as we can simply enforce the
location label of all rule data to match the cluster location.