Default metric metadata to untyped
If now metric metadata can be found in a metadata store, write these
metrics as untyped metrics. This primarily supports metrics created
on the fly by metric relabeling onto the __name__ label.

This is not a first-class use case and merely treated as best-effort.
A warning is logged the first time a metric name is defaulted.