Add ability to pass cert files to operator
Currently, the GMP Operator accepts a Certificate for the webhook server
and a CA Cert for clients as arguments on the command line. This is not
the most secure way of passing the server certificate, as noted by
GKE Security (b/297043491#comment2).

This change allows these certificates to be passed as files instead of
command line arguments and prioritizes the files over the command line
args (args will be ignored if files exist).

Expected usage is that a Kubernetes Secret will be created with the server cert and
a ConfigMap will be created with the Client CA, which can then be
mounted to the operator as a projected volume at the default path of
`/etc/tls/private/`, using standard filenames `tls.crt`, `tls.key`, and
`ca.crt`, respectively.

The paths can be customized using the new flags, if desired. These flags
to the operator are `--cert-file`, `--key-file`, and `--client-ca-file`,
respectively.

The `--tls-cert-base64`, `--tls-key-base64`, and `--ca-cert--base64`
flags should be considered deprecated and are planned to be removed in
the future.