fix: store self-signed certs in secret
Fix issue with operator trying to write self-signed certificates to a
read-only directory by creating the missing TLS secret, which will then
be mounted to the operator in the same way as GKE-provided certs.

This will only attempt to create the webhook-tls secret if it doesn't
exist because the operator checks for the existence of the files from
the projected volume whose source is the webhook-tls secret. Therefore,
the operator only needs the create permission, and not any other
permissions.