fix: limit size of kind cluster names
Due to https://github.com/kubernetes-sigs/kind/issues/623, we need to
ensure we do not violate the length of a kind cluster's name.

We have been ensuring unique cluster names by using the name of the go
test function as the cluster name. However, for longer functions, we run
into this issue.

This change fixes the length of the name to 32 characters and appends a
random hash to the end of the name to prevent collisions for
similarly-named tests.