refactor: use helm to generate manifests
Helm offers a more elegant solution to some of the tasks around assembling the manifests for the operator.

There are a few key changes in this PR:
- Operator manifests move to a Helm chart in `charts/operator` instead of `cmd/operator/deploy`.
- Rule-evaluator manifests are split into a separate chart at `charts/rule-evaluator`.
- Helm is added to Bingo and pinned at v3.14.0.
- YQ is added to Bingo and pinned at v4.40.7.
- `update_manifests` only updates the yaml files in `manifests/` and `examples`. It does not attempt to update the Helm Charts. It previously updated the source manifests in `cmd/operator/deploy`.
- CRDs were regenerated with updated line wrapping. No actual changes.
- diff test for generated files produces output instead of being silenced.
- Documentation for the operator `cmd/operator/README.md` was updated to use Helm.
- Update Makefile target `updateversions` to modify `values.yaml` and re-run `helm template`.
- Added common labels to all operator resources.
- Added `addlicense` to enforce license files on manifests.

Image versions now only need to be updated in the `values.yaml` files for their charts.

Manifests are sorted by the [standard Helm ordering](https://helm.sh/docs/intro/using_helm/#:~:text=Helm%20installs%20resources%20in%20the%20following%20order%3A). No changes were ultimately made to the substance of any resource in `manifests/operator.yaml`. This maintains compatibility with GKE tooling, which relies on that manifest.

This change creates an opportunity to use Helm to generate manifests on the GKE side using Helm as well, which could significantly reduce the complexity related to creating GKE-specific manifests.