chore: fix dependabot commit scopes
Basically, dependabot tries to be smart. It detects the previous commit
message, and then extrapolates based on that to determine how to to
create the next one.

With `include: scope` on, we should not include the scope in the prefix,
because [it will be included again](https://github.com/dependabot/dependabot-core/blob/main/common/lib/dependabot/pull_request_creator/pr_name_prefixer.rb#L110).