feat: configurable vpa
Create scaling options on the OperatorConfig, with the ability for
the operator to deploy Vertical Pod Autoscaling (VPA).

Among other benefits, this change allows customers on Autopilot to use
VPA to attempt to exceed the static resource limits. In GKE Standard,
Vertical Pod Autoscaling must be enabled at the cluster level before it
can be enabled by the Operator.