fix(export): add support in onGCE + no conn to metadata case
Fixes b/344740239 (edge case with GKE Metadata Server and GKE sandbox).

* Added debug logging.
* Updated metadata deps to get https://github.com/googleapis/google-cloud-go/pull/9733 & use timeout-ed context
* Moved risky logic from FromFlags, see code comment why.
* Added regession test.

### Alternatives

Everything we do in FromFlags or constructor is within readines period. We
could consider moving potentially "slow" things on slow network or metadata srv
to exporter.Run. This could be questionable as for GMP to work
we at end need export functionality to work, so delaying that information or
making it surface in separation to readiness might not be helpful.

Signed-off-by: bwplotka <bwplotka@google.com>