ci: pause release-please automation experiment
As discussed with @bernot-dev we could remove release-please for
now as we found some potential blockers and we don't have time to
continue the experimentation or invest in forking/updating release-please.

We have some idea how to do all with script (alternative), so removing this
for now as we don't use release-please. We can get all back from git history.

And we get something valuable out of this: Ideas and.... our beloved conform lint
(jokes aside I think everyone (except Lee) learnt how to use it and it gave
us some pretty clear commit history (rdy for changelog automation!).

Signed-off-by: bwplotka <bwplotka@google.com>