feat: add more example self-podmonitorings
The gmp-operator exposes Prometheus metrics that can be helpful to debug
issues with managed-collection. This is also the case with the managed
alertmanager.

We have examples of how to scrape metrics from other components, but
not the operator or the alertmanager.

So here we provide examples to supplement the self-monitoring exporter
documented in
https://cloud.google.com/stackdriver/docs/managed-prometheus/exporters/prometheus?hl=en.

Partly addresses #793.

Signed-off-by: Danny Clark <danielclark@google.com>