feat: prometheus 2.53.4 support
We do this to prepare a release minor that supports the latest Prometheus 2.54
release.

We had to fix some merge conflicts due to upstream refactors; much of that work
was captured by cherry-picking in commit c3d31f9c419b376844c61635a259f1b15c7b1c12.

Additionally we needed to update the Prometheus config tests to check for the
existence of some new fields that were introduced (e.g. http_headers).

We also remove the replace directive in the go.mod file for the older prometheus
and x/exp packages and pin them to prometheus 2.53.4.